update delete to soft delete

+++ src/main/java/com/caochung/recruitment/service/impl/JobServiceImpl.java
@@ -1,6 +1,8 @@
 package com.caochung.recruitment.service.impl;
 
 import com.caochung.recruitment.constant.ErrorCode;
+import com.caochung.recruitment.constant.JobStatusEnum;
+import com.caochung.recruitment.constant.ResumeStatusEnum;
 import com.caochung.recruitment.domain.Job;
 import com.caochung.recruitment.domain.User;
 import com.caochung.recruitment.dto.request.JobRequestDTO;
@@ -8,6 +10,7 @@ import com.caochung.recruitment.dto.response.JobResponseDTO;
 import com.caochung.recruitment.dto.response.PaginationResponseDTO;
 import com.caochung.recruitment.exception.AppException;
 import com.caochung.recruitment.repository.JobRepository;
+import com.caochung.recruitment.repository.ResumeRepository;
 import com.caochung.recruitment.repository.UserRepository;
 import com.caochung.recruitment.service.JobService;
 import com.caochung.recruitment.service.mapper.JobMapper;
@@ -18,17 +21,21 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class JobServiceImpl implements JobService {
     private final JobRepository jobRepository;
     private final UserRepository userRepository;
+    private final ResumeRepository resumeRepository;
     private final JobMapper jobMapper;
 
     @Override
+    @Transactional
     public JobResponseDTO createJob(JobRequestDTO jobRequestDTO) {
         if(this.jobRepository.existsByName(jobRequestDTO.getName())){
             throw new AppException(ErrorCode.JOB_EXISTED);
@@ -67,18 +74,22 @@ public class JobServiceImpl implements JobService {
     }
 
     @Override
+    @Transactional
     public void updateJob(Long id, JobRequestDTO jobRequestDTO) {
         Job job = this.jobRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.JOB_NOT_FOUND));
         if(!job.getName().equals(jobRequestDTO.getName()) && this.jobRepository.existsByName(jobRequestDTO.getName())){
             throw new AppException(ErrorCode.JOB_EXISTED);
         }
         this.jobMapper.fromUpdateJob(jobRequestDTO, job);
-        this.jobRepository.save(job);
     }
 
     @Override
+    @Transactional
     public void deleteJob(Long id) {
         Job job = this.jobRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.JOB_NOT_FOUND));
+        if(this.resumeRepository.existsByJob_IdAndStatusIn(id, List.of(ResumeStatusEnum.PENDING, ResumeStatusEnum.REVIEWING))){
+            throw new AppException(ErrorCode.JOB_HAS_ACTIVE_RESUMES);
+        }
         this.jobRepository.delete(job);
     }
 }
